+++ src/parser.rs
@@ -36,6 +36,7 @@ pub(crate) mod java;
 pub(crate) mod js;
 pub(crate) mod julia;
 pub(crate) mod kotlin;
+pub(crate) mod lua;
 pub(crate) mod objective_c;
 pub(crate) mod python;
 pub(crate) mod rust;