+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Allow users to manage the downloading of WASM grammars not bundled with treetags
 - Add basic tags support for Kotlin
 - Add extension fields support for Elixir
 - Add extension fields support for Julia
+- Add extension fields support for Lua
 
 ### Changed
 - Tag Address is truncated to 96 chars for native parsers